// src/app/pages/contact/contact.d.ts
export type TContactFormData = {
    name: string
    email: string
    message: string
}

export type TSocialNetwork = {
    icon: string
    link: "https://github.com/willfynch" | "https://www.linkedin.com/in/williamfinzy/"
    name: string
}

export type TEmailAlertStatus = "pending" | "error" | "success" | "hidden"